# backend/api-docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Plataforma Scan API Documentation",
    "description": "API documentation for the Plataforma Scan system that provides information for both mobile and web applications of the software.",
    "version": "3.5.0",
    "contact": {
      "name": "Plataforma Scan Support",
      "email": "support@plataformascan.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development Server"
    },
    {
      "url": "https://api.plataformascan.com",
      "description": "Production Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "MANAGER",
              "OPERATOR"
            ]
          },
          "isPending": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "Register a new user",
        "description": "Register a new user with the system",
        "tags": [
          "Authentication"
        ],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password",
                  "name",
                  "verificationCode"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "User's email address"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "description": "User's password"
                  },
                  "name": {
                    "type": "string",
                    "description": "User's full name"
                  },
                  "verificationCode": {
                    "type": "string",
                    "description": "Verification code sent to user's email"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or verification code"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login to the system",
        "description": "Authenticate a user and return a JWT token",
        "tags": [
          "Authentication"
        ],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "User's email address"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "description": "User's password"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "description": "JWT token for authentication"
                    },
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/auth/send-verification-code": {
      "post": {
        "summary": "Send verification code",
        "description": "Sends a verification code to the user's email for registration",
        "tags": [
          "Authentication"
        ],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "User's email address"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification code sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Código de verificação enviado para o seu email"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Email already in use"
          }
        }
      }
    },
    "/auth/generate-verification-code": {
      "post": {
        "summary": "Generate verification code for admin",
        "description": "Generates a verification code for admin to create manager accounts",
        "tags": [
          "Authentication"
        ],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Manager's email address"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification code generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Código de verificação gerado com sucesso"
                    },
                    "verificationCode": {
                      "type": "string",
                      "example": "a1b2c3"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Email already in use by an active user"
          }
        }
      }
    },
    "/auth/password-recovery/request": {
      "post": {
        "summary": "Request password reset",
        "description": "Sends a password reset code to the user's email",
        "tags": [
          "Authentication"
        ],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "User's email address"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reset code sent successfully"
          },
          "400": {
            "description": "Invalid email or user not found"
          }
        }
      }
    },
    "/auth/password-recovery/verify-code": {
      "post": {
        "summary": "Verify password reset code",
        "description": "Verifies the reset code sent to the user's email",
        "tags": [
          "Authentication"
        ],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "resetCode"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "User's email address"
                  },
                  "resetCode": {
                    "type": "string",
                    "description": "Reset code sent to user's email"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reset code verified successfully"
          },
          "400": {
            "description": "Invalid or expired reset code"
          }
        }
      }
    },
    "/auth/password-recovery/reset": {
      "post": {
        "summary": "Reset password",
        "description": "Resets the user's password with a new one",
        "tags": [
          "Authentication"
        ],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "resetCode",
                  "newPassword"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "User's email address"
                  },
                  "resetCode": {
                    "type": "string",
                    "description": "Reset code sent to user's email"
                  },
                  "newPassword": {
                    "type": "string",
                    "format": "password",
                    "description": "New password"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset successfully"
          },
          "400": {
            "description": "Invalid or expired reset code, or missing required fields"
          }
        }
      }
    },
    "/auth/verify-token": {
      "get": {
        "summary": "Verify JWT token",
        "description": "Verifies if the JWT token is valid",
        "tags": [
          "Authentication"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Token is valid",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "valid": {
                      "type": "boolean",
                      "example": true
                    },
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid token"
          }
        }
      }
    },
...
  }
}